// src/interfaces/model/limitObject.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Load control limits
 */
export interface LimitObject {
  /**
   * The number of seconds that the rate window should encompass
   */
  period: number
  /**
   * The number of requests that are allowed in the specified rate limiting window
   */
  rate: number
}