// src/interfaces/model/quotaObject.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Load control quota
 */
export interface QuotaObject {
  /**
   * The maximum number of requests allowed during the quota period
   */
  quota: number
  /**
   *          The time, in seconds. during which the quota is valid. So for 1000 requests per hour, this          value would be 3600 while quota_max and quota_remaining would be 1000.
   */
  quotaRenewalRate: number
  /**
   * An epoch that defines when the quota renews
   */
  quotaRenews: number
}